split mustache into smaller modules
