+++ project-words.txt
@@ -1,18 +1,19 @@
 btreemap
 entr
-insta
-jsonplaceholder
-maplit
-mdast
-opentelemetry
-peekable
-tailcall
-tailcallrc
-thiserror
 Hasher
 Htpasswd
 httpmock
 inlines
+insta
+jsonplaceholder
 Jwks
+maplit
+mdast
+multispace
+opentelemetry
+peekable
 schemars
+tailcall
+tailcallrc
+thiserror
 tupled
