+++ src/core/mod.rs
@@ -47,6 +47,7 @@ use std::num::NonZeroU64;
 use async_graphql_value::ConstValue;
 use http::Response;
 use ir::model::IoId;
+pub use mustache::Mustache;
 pub use tailcall_macros as macros;
 
 pub trait EnvIO: Send + Sync + 'static {
