+++ src/core/mustache/mod.rs
@@ -1,0 +1,5 @@
+mod model;
+mod parse;
+mod render;
+
+pub use model::*;
